+++ angular-workspace/projects/ni/nimble-angular/src/directives/dialog/nimble-dialog.directive.ts
@@ -23,8 +23,6 @@ export class NimbleDialogDirective<CloseReason = void> {
         return this.elementRef.nativeElement.preventDismiss;
     }
 
-    // preventDismiss property intentionally maps to the prevent-dismiss attribute
-    // eslint-disable-next-line @angular-eslint/no-input-rename
     @Input('prevent-dismiss') public set preventDismiss(value: BooleanValueOrAttribute) {
         this.renderer.setProperty(this.elementRef.nativeElement, 'preventDismiss', toBooleanProperty(value));
     }
@@ -33,8 +31,6 @@ export class NimbleDialogDirective<CloseReason = void> {
         return this.elementRef.nativeElement.headerHidden;
     }
 
-    // Renaming because property should have camel casing, but attribute should not
-    // eslint-disable-next-line @angular-eslint/no-input-rename
     @Input('header-hidden') public set headerHidden(value: BooleanValueOrAttribute) {
         this.renderer.setProperty(this.elementRef.nativeElement, 'headerHidden', toBooleanProperty(value));
     }
@@ -43,8 +39,6 @@ export class NimbleDialogDirective<CloseReason = void> {
         return this.elementRef.nativeElement.footerHidden;
     }
 
-    // Renaming because property should have camel casing, but attribute should not
-    // eslint-disable-next-line @angular-eslint/no-input-rename
     @Input('footer-hidden') public set footerHidden(value: BooleanValueOrAttribute) {
         this.renderer.setProperty(this.elementRef.nativeElement, 'footerHidden', toBooleanProperty(value));
     }